requirements.txt
python-binance
pandas
python-telegram-bot==13.7
ta
requests
